Move multiplayer screen components to osu.Game/Screens/Multi/Components
